refactor(fe): 튜토리얼 데이터 경로 수정
일관성을 위해 data/ -> const/ 로 수정
