+++ frontend/src/pages/tutorial/const/tutorialBattle.ts
